slicing movie_detail_page
- add movie_detail_page route
- fix ontap in movie_item widget
- in movie_page, change refresh indicator placement, now in the head of body, and add a padding when getMovieListError, and then fill the onTap value with goes to movieDetailPage
- add dependencies hydrated_bloc: ^9.1.5 and path_provider: ^2.1.4, but not used yet.

in this movie_detail_page, i use data from movie_page so i dont need to get the data again to .themoviedb cause time efficiency. but if we want to show more data detail that we cant find it in movieModel, we can use movieId to get it from .themoviedb api.